Fix moving const apiColorMap to backend to fix build problems
