+++ src/app/lib/colors.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import type { apiColorMap } from '../../lib/types/image';
 
 export type color = {
   cssVar: string;
@@ -105,74 +106,5 @@ export const COLORS = {
     gradientFillClass: `fillInactiveGradient`,
     gradientStrokeClass: `strokeInactiveGradient`,
     gradientSVG: GRADIENTS_SVG.inactive,
-  },
-};
-
-export const apiColorMap = {
-  monochrom: {
-    unsplash: 'black_and_white',
-    // pexels: 'black',
-    pexels: 'noRequest',
-    pixabay: 'grayscale',
-  },
-  black: {
-    unsplash: 'black',
-    // pexels: 'black',
-    pexels: 'noRequest',
-    pixabay: 'black',
-  },
-  white: {
-    unsplash: 'white',
-    // pexels: 'white',
-    pexels: 'noRequest',
-    pixabay: 'white',
-  },
-  gray: {
-    unsplash: 'noRequest',
-    // pexels: 'gray',
-    pexels: 'noRequest',
-    pixabay: 'gray',
-  },
-  red: {
-    unsplash: 'red',
-    // pexels: 'red',
-    pexels: 'noRequest',
-    pixabay: 'red',
-  },
-  orange: {
-    unsplash: 'orange',
-    // pexels: '#e68600',
-    pexels: 'noRequest',
-    pixabay: 'orange,brown',
-  },
-  yellow: {
-    unsplash: 'yellow',
-    // pexels: 'yellow',
-    pexels: 'noRequest',
-    pixabay: 'yellow',
-  },
-  green: {
-    unsplash: 'green',
-    // pexels: 'green',
-    pexels: 'noRequest',
-    pixabay: 'green',
-  },
-  teal: {
-    unsplash: 'teal',
-    // pexels: 'turquoise',
-    pexels: 'noRequest',
-    pixabay: 'turquoise',
-  },
-  blue: {
-    unsplash: 'blue',
-    // pexels: 'blue',
-    pexels: 'noRequest',
-    pixabay: 'blue',
-  },
-  purple: {
-    unsplash: 'purple',
-    // pexels: '#b366ff',
-    pexels: 'noRequest',
-    pixabay: 'lilac,pink',
   },
 };
